reorganization of io.py and simulations.py
